Register all token standards in a single file
Collapse the per-standard files (ethereum/collateral/dtcc) into one
register.ts: registerTokenStandards plus the three per-group registrars
live together, index.ts re-exports. No behavior change.
