+++ src/integrations/token-standards/index.ts
@@ -1,4 +1,1 @@
-export * from "./ethereum";
-export * from "./collateral";
-export * from "./dtcc";
 export * from "./register";
